Move debug assets to playground
